init/generate_data/requirements.txt
# 数据生成工具依赖
psycopg2-binary>=2.9.9
